requirements.txt
appnope==0.1.0
backcall==0.1.0
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
decorator==4.4.0
Django==2.2.4
django-debug-toolbar==2.0
django-filter==2.2.0
djangorestframework==3.10.2
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.15.1
Markdown==3.1.1
newsapi-python==0.2.5
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py-bcrypt==0.4
pycparser==2.19
Pygments==2.4.2
pytz==2019.2
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
